requirements.txt
pywsse
requests
Flask
flask_talisman
python-dateutil
Jinja2
